+++ Control/Joint/Effects.hs
@@ -3,4 +3,4 @@ module Control.Joint.Effects (module Exports) where
 import Control.Joint.Effects.State as Exports
 import Control.Joint.Effects.Maybe as Exports
 import Control.Joint.Effects.Either as Exports
-import Control.Joint.Effects.Configured as Exports
+import Control.Joint.Effects.Reader as Exports